Fix for issue #54 where branches (and tags) with '/' in the name could not be viewed
Use the PathEncoder class (introduced to fix issue #55) to encode branch and tag names (in addition to repository paths) when used with the Tree/Blob/Raw/Commits controllers. All non-RFC 3986 unreserved characters in branch and tag names are encoded (in a manner similar to URL/percent encoding) and will not be mis-interpreted by IIS as URL components.

changelog.md is updated to describe the fix.

A few instances of trailing whitespace and two unnecessary @usings were removed.