Added a Readme.md field for Repositories for a readme page using Markdown
Very similar to the readme.md page for github pages. For each repository
there is a new Readme property that is a string of markdown text that
describes the repository.  This could be used to describe the
dependencies, instructions and other information about the repository.
Added the MarkdownDeep nuget package to handle the transformation of
markdown text to Html.  The user interface includes a markdown toolbar,
help section and preview fields.