Fix LDAP authentication issues on push
The user should be able to use a shortened domain name (e.g. contoso\username instead of contoso.com\username) when logging on to the server via NTLM/LDAP. This works fine for the web interface, but produces 401 errors when trying to push changes to Git. We already have a setting for the full domain name in the web.config, which is used in other parts of the system. There's no reason to try and guess the domain name by taking apart the username.