Performance improvements via caching of PrincipalSearchResults
This commit provides huge performance gains by caching the results of
calls to GroupPrincipal.GetMembers(true) by enumerating the
PrincipalSearchResults object and storing it. Normally this class lazy
loads the results, so the enum is triggered up front and stored for
later use. On my use case this sped up authorization calls from 10+
seconds to under a second.

Like the previous commit, the cache is set to expire after 10 minutes -
I will endeavour to move this out into web.config.